Add a debian_src setup command, and add a version.py file
version.py provides the version number in various formats, for packaging purposes. This does not change the fact that the unique place for changing MAJOR, MINOR, MICRO is src/fisx_version.h.

Currently, version.py is not included in the build (unlike silx). It is only imported in setup.py and build-deb.sh.